+++ src/components/tests/PivotTable.spec.tsx
@@ -186,17 +186,4 @@ describe("PivotTable", () => {
         expect(wrapper.prop("afm")).toEqual(expectedAfm);
         expect(wrapper.prop("resultSpec")).toEqual(expectedResultSpec);
     });
-
-    describe("groupRows for attribute columns", () => {
-        it.each([["should", true, true], ["should NOT", undefined, false], ["should NOT", false, false]])(
-            "%s group rows for attribute columns when groupRows is %s",
-            (_should, groupRows, expected) => {
-                const component = renderShallowComponent({
-                    groupRows,
-                });
-
-                expect(component.prop("groupRows")).toEqual(expected);
-            },
-        );
-    });
 });